Minor refactoring in examples
