+++ examples/inverted_pendulum.py
@@ -7,33 +7,14 @@ import tensorflow as tf
 from tensorflow import logging
 np.random.seed(0)
 
-
-def rollout(policy, timesteps):
-    X = []; Y = []
-    env.reset()
-    x, _, _, _ = env.step(0)
-    for timestep in range(timesteps):
-        env.render()
-        u = policy(x)
-        x_new, _, done, _ = env.step(u)
-        if done: break
-        X.append(np.hstack((x, u)))
-        Y.append(x_new - x)
-        x = x_new
-    return np.stack(X), np.stack(Y)
-
-def random_policy(x):
-    return env.action_space.sample()
-
-def pilco_policy(x):
-    return pilco.compute_action(x[None, :])[0, :]
+from utils import rollout, policy
 
 with tf.Session(graph=tf.Graph()) as sess:
     env = gym.make('InvertedPendulum-v2')
     # Initial random rollouts to generate a dataset
-    X,Y = rollout(policy=random_policy, timesteps=40)
+    X,Y = rollout(env=env, pilco=None, random=True, timesteps=40)
     for i in range(1,3):
-        X_, Y_ = rollout(policy=random_policy, timesteps=40)
+        X_, Y_ = rollout(env=env, pilco=None, random=True,  timesteps=40)
         X = np.vstack((X, X_))
         Y = np.vstack((Y, Y_))
 
@@ -56,7 +37,7 @@ with tf.Session(graph=tf.Graph()) as sess:
         pilco.optimize_models()
         pilco.optimize_policy()
         import pdb; pdb.set_trace()
-        X_new, Y_new = rollout(policy=pilco_policy, timesteps=100)
+        X_new, Y_new = rollout(env=env, pilco=pilco, timesteps=100)
         print("No of ops:", len(tf.get_default_graph().get_operations()))
         # Update dataset
         X = np.vstack((X, X_new)); Y = np.vstack((Y, Y_new))
